updated example code to be a bit cleaner
