+++ distbelief/server.py
@@ -7,9 +7,8 @@ import torch.distributed as dist
 from distbelief.utils.messaging import MessageCode, MessageListener, send_message
 from distbelief.utils.serialization import ravel_model_params
 
-logging.basicConfig(level=logging.INFO)
-
 _LOGGER = logging.getLogger(__name__)
+_LOGGER.setLevel(logging.INFO)
 
 class ParameterServer(MessageListener):
     """ParameterServer"""
@@ -20,7 +19,7 @@ class ParameterServer(MessageListener):
         super().__init__(model)
 
     def receive(self, sender, message_code, parameter):
-        _LOGGER.info("Processing message: {} from sender {}".format(message_code.name, sender))
+        print("Processing message: {} from sender {}".format(message_code.name, sender))
 
         if message_code == MessageCode.ParameterUpdate:
             #be sure to clone here
