+++ distbelief/utils/distributed.py
@@ -7,11 +7,11 @@ import torch.distributed as dist
 
 _LOGGER = logging.getLogger(__name__)
 
-def init_processes(rank, size, fn, backend='tcp'):
+def init_processes(args, fn, backend='tcp'):
     """ Initialize the distributed environment.
     Server and clients must call this as an entry point.
     """
     os.environ['MASTER_ADDR'] = 'localhost'
     os.environ['MASTER_PORT'] = '29500'
-    dist.init_process_group(backend, rank=rank, world_size=size)
-    fn()
+    dist.init_process_group(backend, rank=args.rank, world_size=args.world_size)
+    fn(args)
