+++ example/main.py
@@ -1,3 +1,5 @@
+import logging 
+import argparse
 import torch
 import torchvision
 import torchvision.transforms as transforms
@@ -5,91 +7,14 @@ import numpy as np
 import torch.nn as nn
 import torch.nn.functional as F
 
+from models import LeNet, AlexNet
+
 import torch.optim as optim
 from distbelief.optim import DownpourSGD
-import threading
-import argparse
-import logging
+from distbelief.utils.distributed import init_processes
+from distbelief.server import ParameterServer
 
-# logging.basicConfig(level=logging.INFO)
-# _LOGGER = logging.getLogger(__name__)
-
-class LeNet(nn.Module):
-    def __init__(self):
-        super(Net, self).__init__()
-        self.conv1 = nn.Conv2d(3, 6, kernel_size=5)
-        self.conv2 = nn.Conv2d(6, 16, kernel_size=5)
-        self.conv2_drop = nn.Dropout2d()
-        self.fc1 = nn.Linear(16 * 5 * 5, 120)
-        self.fc2 = nn.Linear(120, 84)
-        self.fc3 = nn.Linear(84, 10)
-
-    def forward(self, x):
-        x = F.relu(F.max_pool2d(self.conv1(x), 2))
-        x = F.relu(F.max_pool2d(self.conv2_drop(self.conv2(x)), 2))
-        x = x.view(-1, 16* 5* 5)
-        x = F.relu(self.fc1(x))
-        x = F.dropout(x, training=self.training)
-        x = F.relu(self.fc2(x))
-        x = self.fc3(x)
-        return x
-
-class AlexNet(nn.Module):
-    def __init__(self, num_classes=10):
-        super(AlexNet, self).__init__()
-        self.features = nn.Sequential(
-            nn.Conv2d(3, 64, kernel_size=11, stride=4, padding=5),
-            nn.ReLU(inplace=True),
-            nn.MaxPool2d(kernel_size=2, stride=2),
-            nn.Conv2d(64, 192, kernel_size=5, padding=2),
-            nn.ReLU(inplace=True),
-            nn.MaxPool2d(kernel_size=2, stride=2),
-            nn.Conv2d(192, 384, kernel_size=3, padding=1),
-            nn.ReLU(inplace=True),
-            nn.Conv2d(384, 256, kernel_size=3, padding=1),
-            nn.ReLU(inplace=True),
-            nn.Conv2d(256, 256, kernel_size=3, padding=1),
-            nn.ReLU(inplace=True),
-            nn.MaxPool2d(kernel_size=2, stride=2),
-        )
-        self.classifier = nn.Linear(256, num_classes)
-
-    def forward(self, x):
-        x = self.features(x)
-        x = x.view(x.size(0), -1)
-        x = self.classifier(x)
-        return x
-
-flag_use_cuda = False
-
-def main(*args, **kwargs):
-    parser = argparse.ArgumentParser(description='Distbelief training example')
-    parser.add_argument('--batch-size', type=int, default=64, metavar='N', help='input batch size for training (default: 64)')
-    parser.add_argument('--test-batch-size', type=int, default=1000, metavar='N', help='input batch size for testing (default: 1000)')
-    parser.add_argument('--epochs', type=int, default=10, metavar='N', help='number of epochs to train (default: 10)')
-    parser.add_argument('--lr', type=float, default=0.01, metavar='LR', help='learning rate (default: 0.01)')
-    parser.add_argument('--momentum', type=float, default=0.5, metavar='M', help='SGD momentum (default: 0.5)')
-    parser.add_argument('--no-cuda', action='store_true', default=False, help='disables CUDA training')
-    parser.add_argument('--seed', type=int, default=1, metavar='S', help='random seed (default: 1)')
-    parser.add_argument('--log-interval', type=int, default=10, metavar='N', help='how many batches to wait before logging training status')
-    parser.add_argument('--dataset', type=str, default='MNIST', help='Dataset to train with (CIFAR or MNIST)')
-    args = parser.parse_args()
-
-    # if args.dataset == 'MNIST':
-    #     trainloader = torch.utils.data.DataLoader(
-    #     torchvision.datasets.MNIST('../data', train=True, download=True,
-    #                    transform=transforms.Compose([
-    #                        transforms.ToTensor(),
-    #                        transforms.Normalize((0.1307,), (0.3081,))
-    #                    ])),
-    #     batch_size=args.batch_size, shuffle=True, **kwargs)
-    #     testloader = torch.utils.data.DataLoader(
-    #         torchvision.datasets.MNIST('../data', train=False, transform=transforms.Compose([
-    #                            transforms.ToTensor(),
-    #                            transforms.Normalize((0.1307,), (0.3081,))
-    #                        ])),
-    #         batch_size=args.test_batch_size, shuffle=True, **kwargs)
-    # TODO: distinguishing between MNIST and CIFAR for experimentation
+def main(args):
 
     transform = transforms.Compose([
                 transforms.ToTensor(),
@@ -97,31 +22,31 @@ def main(*args, **kwargs):
             ])
 
     trainset = torchvision.datasets.CIFAR10(root='./data', train=True, download=True, transform=transform)
-    trainloader = torch.utils.data.DataLoader(trainset, batch_size=64, shuffle=True, num_workers=1)
+    trainloader = torch.utils.data.DataLoader(trainset, batch_size=args.batch_size, shuffle=True, num_workers=2)
 
-    testset = torchvision.datasets.CIFAR10(root='../data', train=False, download=True, transform=transform)
-    testloader = torch.utils.data.DataLoader(testset, batch_size=64, shuffle=False, num_workers=1)
+    testset = torchvision.datasets.CIFAR10(root='./data', train=False, download=True, transform=transform)
+    testloader = torch.utils.data.DataLoader(testset, batch_size=args.test_batch_size, shuffle=False, num_workers=2)
 
-    classes = ('plane', 'car', 'bird', 'cat', 'deer', 'dog', 'frog', 'horse', 'ship', 'truck')
+    
 
     net = AlexNet()
 
-    optimizer = DownpourSGD(net.parameters(), lr=0.1, freq=10, model=net)
+    optimizer = DownpourSGD(net.parameters(), lr=args.lr, freq=args.freq, model=net)
     # optimizer = optim.SGD(net.parameters(), lr=0.1, momentum=0.0)
     scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer, patience=1, verbose=True, min_lr=1e-3)
 
+    # train
     net.train()
-    if flag_use_cuda:
+    if args.cuda:
         net = net.cuda()
 
-    num_print = 20
-    for epoch in range(25):  # loop over the dataset multiple times
+    for epoch in range(args.epochs):  # loop over the dataset multiple times
         running_loss = 0.0
         for i, data in enumerate(trainloader, 0):
             # get the inputs
             inputs, labels = data
 
-            if flag_use_cuda:
+            if args.cuda:
                 inputs, labels = inputs.cuda(), labels.cuda()
 
             # zero the parameter gradients
@@ -135,25 +60,28 @@ def main(*args, **kwargs):
 
             # print statistics
             running_loss += loss.item()
-            if i % num_print == 0:    # print every n mini-batches
-                print('Epoch: %d, Iteration: %5d loss: %.3f' % (epoch, i, running_loss / num_print))
+            if i % args.log_interval == 0 and i > 0:    # print every n mini-batches
+                print('Epoch: %d, Iteration: %5d loss: %.3f' % (epoch, i, running_loss / args.log_interval))
                 running_loss = 0.0
-        val_loss = evaluate(net, testloader, classes)
+
+        val_loss = evaluate(net, testloader, args)
         scheduler.step(val_loss)
 
     print('Finished Training')
 
 
-def evaluate(net, testloader, classes):
+def evaluate(net, testloader, args):
+    classes = ('plane', 'car', 'bird', 'cat', 'deer', 'dog', 'frog', 'horse', 'ship', 'truck')
     net.eval()
     running_loss = 0.0
     class_correct = list(0. for i in range(10))
     class_total = list(0. for i in range(10))
+
     with torch.no_grad():
         for data in testloader:
             images, labels = data
 
-            if flag_use_cuda:
+            if args.cuda:
                 images, labels = images.cuda(), labels.cuda()
 
             outputs = net(images)
@@ -175,5 +103,26 @@ def evaluate(net, testloader, classes):
     
     return running_loss
 
-if __name__ == '__main__':
-    main()
+def init_server(args):
+    model = AlexNet()
+    server = ParameterServer(model=model)
+    server.start()
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description='Distbelief training example')
+    parser.add_argument('--batch-size', type=int, default=64, metavar='N', help='input batch size for training (default: 64)')
+    parser.add_argument('--test-batch-size', type=int, default=1000, metavar='N', help='input batch size for testing (default: 1000)')
+    parser.add_argument('--epochs', type=int, default=10, metavar='N', help='number of epochs to train (default: 10)')
+    parser.add_argument('--lr', type=float, default=0.1, metavar='LR', help='learning rate (default: 0.1)')
+    parser.add_argument('--freq', type=int, default=10, metavar='N', help='how often to send/pull grads (default: 10)')
+    parser.add_argument('--cuda', action='store_true', default=False, help='use CUDA for training')
+    parser.add_argument('--log-interval', type=int, default=10, metavar='N', help='how many batches to wait before logging training status')
+    parser.add_argument('--rank', type=int, metavar='N', help='rank of current process (0 is server, 1+ is training node)')
+    parser.add_argument('--world-size', type=int, default=3, metavar='N', help='size of the world')
+    parser.add_argument('--server', action='store_true', default=False, help='server node?')
+
+    args = parser.parse_args()
+    if args.server:
+        init_processes(args, init_server)
+    else:
+        init_processes(args, main)
